+++ tests/test_routes.py
@@ -52,7 +52,9 @@ def test_api_get_data_as_json(client, populated_db):
     rv = client.get("/api/v1/data/foo")
     assert rv.status_code == 200
     assert rv.is_json
-    d = rv.get_json()['rows']
+    d = rv.get_json()
+    assert d['units'] == None
+    d = d['rows']
     assert d[0]['n'] == 0
     assert d[0]['value'] == 15
     assert d[3]['value'] == 9