+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 
 ## Unreleased
++ `units` are now also returned in the GET `/data` API call.
 + Removed a confusing route: `/api/v1/<metric>`. (#39)
 + Added a title that will link back to the index page. (#40)
 + Changed the way we handle generating links when behind a proxy that's